% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbib.R
\name{getbib}
\alias{getbib}
\title{Title}
\usage{
getbib(
  file_path = "ref.bib",
  refresh_days = 0,
  copy_to_current = FALSE,
  URL = "http://drsmorey.org/bibtex/bibfile.php",
  convert_from = NULL,
  convert_to = NULL
)
}
\arguments{
\item{file_path}{Where to save the bib file}

\item{refresh_days}{How old the existing file can be before it is refreshed}

\item{copy_to_current}{Copy the file to the current directory for backup?}

\item{URL}{URL of the bib file}

\item{convert_from}{NULL to not use iconv; otherwise string giving encoding}

\item{convert_to}{string giving encoding; only used if `convert_from` is non-null}
}
\value{
Returns a path to the bib file
}
\description{
Title
}
