% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_slide.R
\name{new_slide}
\alias{new_slide}
\title{Create a new slide in the middle of a knitr chunk}
\usage{
new_slide(title = "\\n\\n## New slide\\n\\n")
}
\arguments{
\item{title}{The title of the new slide}
}
\value{
An object a class named knit_asis, so that `knitr`` 
will treat it as is (the effect is the same as the chunk 
option `results = 'asis'`) when it is written to the output.
}
\description{
Create a new slide in the middle of a knitr chunk
}
