% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_options.R
\name{pkg_options}
\alias{pkg_options}
\title{Set or get options for my package}
\usage{
pkg_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported:
\itemize{
 \item{\code{wait}}{(\code{numeric};2) Amount of time in seconds to wait for HTML to render before printing PDF }
 \item{\code{printBackground}}{(\code{logical};TRUE) Print the HTML background images/color to PDF?}
 \item{\code{marginTop}}{(\code{numeric};0) Top margin to add to PDF slides}
 \item{\code{marginBottom}}{(\code{numeric};0) Bottom margin to add to PDF slides}
 \item{\code{marginLeft}}{(\code{numeric};0) Left margin to add to PDF slides}
 \item{\code{marginRight}}{(\code{numeric};0) Right margin to add to PDF slides}
 \item{\code{paperHeight}}{(\code{numeric};11) Width of the slides in inches (height because landscape)}
 \item{\code{aspect}}{(\code{numeric};4/3) Aspect ratio of non-widescreen slides}
 \item{\code{widescreen_aspect}}{(\code{numeric};11/7) Aspect ratio of widescreen slides}
 \item{\code{notes_slides}}{(\code{logical};FALSE) Compile notes slides when using the add-in? }
 \item{\code{pdfjam_offset}}{(\code{character};"-7cm 0cm") Amount to shift to the right and down}
 \item{\code{pdfjam_delta}}{(\code{character};"0cm .5cm") Amount of horizontal and vertical padding}
 \item{\code{pdfjam_scale}}{(\code{character};"0.8") Size by which to scale the slides }
 \item{\code{pdfjam_nup}}{(\code{character};"1x2") Layout (columns by rows, because landscape)}
 \item{\code{pdfjam_frame}}{(\code{character};"true") Add a frame around the slides?}
 \item{\code{pdfjam_noautoscale}}{(\code{character};"false") activate/deactivate autoscaling}
 \item{\code{pdfjam_suffix}}{(\code{character};"notes") Suffix to add to notes pdf file name }
}

For the pdfjam options, see \url{https://warwick.ac.uk/fac/sci/statistics/staff/academic-research/firth/software/pdfjam121/}
and \url{http://texdoc.net/texmf-dist/doc/latex/pdfpages/pdfpages.pdf}.
}

