% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html_and_pdf.R
\name{render_html_and_pdf0}
\alias{render_html_and_pdf0}
\title{Render an Rmarkdown presentation to HTML and PDF}
\usage{
render_html_and_pdf0(rmd_filename, notes_slides = pkg_options()$notes_slides)
}
\arguments{
\item{rmd_filename}{Rmd file to compile}

\item{notes_slides}{Create a PDF with space for notes?}
}
\value{
Returns a list with the output file paths
}
\description{
Render an Rmarkdown presentation to HTML and PDF
}
