% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html_and_pdf.R
\name{render_html_and_pdf}
\alias{render_html_and_pdf}
\title{Render an Rmarkdown presentation to HTML and PDF}
\usage{
render_html_and_pdf()
}
\value{
Returns a list with the output file paths
}
\description{
Render an Rmarkdown presentation to HTML and PDF
}
